requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
python-dotenv>=1.0.0
openai>=1.3.0
